Add translations
The previous incarnation of this mod had a single translation string, which now corresponds to `options.borderlessFullscreen`.
Many users contributed proper translations in their native languages, and I want to ensure their contributions are included in this commit.
However, I had to fill in the newer entries with machine translations, so keep that in mind.

For clarity, here's the list of contributors and their translations, where "MT" stands for "Machine Translation":

 - `ar_sa` (MT)
 - `be_by` (MT)
 - `de_de` (MT)
 - `en_us` (MT)
 - `es_es` (MT)
 - `es_mx` (@TheLegendofSaram, MT)
 - `fr_fr` (MT)
 - `it_it` (MT)
 - `ja_jp` (MT)
 - `ko_kr` (@301haru, MT)
 - `nl_nl` (MT)
 - `pl_pl` (@Radk6, MT)
 - `pt_pt` (MT)
 - `ru_ru` (@mpustovoi, MT)
 - `uk_ua` (@Dessader, @Kir-Antipov)
 - `zh_tw` (@yichifauzi, MT)

---------

Co-authored-by: Santiago Hernandez <73335686+TheLegendofSaram@users.noreply.github.com>
Co-authored-by: 301haru <michael1302@naver.com>
Co-authored-by: Rad <63958709+Radk6@users.noreply.github.com>
Co-authored-by: Korben <92390626+mpustovoi@users.noreply.github.com>
Co-authored-by: Dessader <dessader.live@gmail.com>
Co-authored-by: dirtTW <31981880+yichifauzi@users.noreply.github.com>