Fix config screen registration on Forge
How many fucking times do they need to rename classes that are part of their core API?

Surely, it's not like they manage a massive mod loader that thousands of mods rely on and expect at least some minimal level of stability.

What a fucking shitshow of a mod loader.